requirements.txt
fastapi
uvicorn[standard]
requests
django
python-dotenv
pandas
